Update the service w/ several enhancements
This update adds the first working route that uses SQL queries via the
JDBI.

Unused portions of the service have been removed as well; particularly,
the modules that came pre-packaged with the play-java template.

Most importantly, the sbt-plugin version in projects/plugins.sbt was
downgraded from 2.5.5 to 2.4.3. This took care of a deprecated
play.db.DB module in the newer version that caused the app to crash.